+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Removed `Hyper` variant from `Error` type
 - Bumped `http-endpoint` dependency to `0.2`
   - Introduced new `RequestError` type
 - Bumped `websocket-util` dependency to `0.4`